duck working fds merda do uncreal

+++ Source/YellPointAndPray/Obstacles/Guard/Guard.h
@@ -51,7 +51,7 @@ protected:
 
 	int SuspicionMax = 100;
 	UPROPERTY(Replicated)
-	int CurrentSuspicion = 0;
+	float CurrentSuspicion = 0;
 
 	UFUNCTION()
 	void Patrol();
